Fix test sequence shared by two types of history
Both Browser and History shares the first step of the ReplaceChangesTheKey sequence.
The expectation to match the key prop is different for Memory since it creates
a first initial entry which contains a value other than undefined. Checking if
entries is part of the returned history objects is now used to distinguish one history
object type from another and write a working expecation match object.